requirements.txt
pandas
biopython
nltk